Improve JSON schemata
This change is mostly about more finegrained validations: e.g., only
require a "scale" property to be a number when used with the decimal
logical type, and validate the underlying type for logical types (this
last bit includes changes introduced in Avro 1.12.0).

As a bonus, it also reduces line count and documents its rather odd
structure to work around two JetBrains bugs.